Refactor broker_manager to carve out pksbroker_manager and vcdbroker_manager.
